// services/webapp/src/sdk/models/ClubRow.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Open Dossard
 * Documentation de l\'API Open Dossard
 *
 * The version of the OpenAPI document: 1.0
 * Contact: contact@opendossard.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ClubRow
 */
export interface ClubRow {
    /**
     * 
     * @type {number}
     * @memberof ClubRow
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ClubRow
     */
    longName?: string;
    /**
     * 
     * @type {string}
     * @memberof ClubRow
     */
    dept?: string;
    /**
     * 
     * @type {string}
     * @memberof ClubRow
     */
    shortName?: string;
}

export function ClubRowFromJSON(json: any): ClubRow {
    return ClubRowFromJSONTyped(json, false);
}

export function ClubRowFromJSONTyped(json: any, ignoreDiscriminator: boolean): ClubRow {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'longName': !exists(json, 'longName') ? undefined : json['longName'],
        'dept': !exists(json, 'dept') ? undefined : json['dept'],
        'shortName': !exists(json, 'shortName') ? undefined : json['shortName'],
    };
}

export function ClubRowToJSON(value?: ClubRow | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'longName': value.longName,
        'dept': value.dept,
        'shortName': value.shortName,
    };
}